+++ oshi-core/src/main/java/oshi/hardware/platform/windows/WindowsBaseboardJNA.java
@@ -4,67 +4,21 @@
  */
 package oshi.hardware.platform.windows;
 
-import static oshi.util.Memoizer.memoize;
-
-import java.util.function.Supplier;
+import java.util.Objects;
 
 import oshi.annotation.concurrent.Immutable;
-import oshi.driver.common.windows.wmi.Win32BaseBoard.BaseBoardProperty;
-import oshi.driver.common.windows.wmi.WmiResult;
-import oshi.driver.common.windows.wmi.WmiUtil;
-import oshi.driver.windows.wmi.Win32BaseBoardJNA;
-import oshi.hardware.common.AbstractBaseboard;
-import oshi.util.Constants;
-import oshi.util.Util;
-import oshi.util.tuples.Quartet;
+import oshi.driver.common.windows.wmi.WmiQueryExecutor;
+import oshi.hardware.common.platform.windows.WindowsBaseboard;
+import oshi.util.platform.windows.WmiQueryExecutorJNA;
 
 /**
- * Baseboard data obtained from WMI
+ * Baseboard data obtained from WMI using JNA.
  */
 @Immutable
-final class WindowsBaseboardJNA extends AbstractBaseboard {
-
-    private final Supplier<Quartet<String, String, String, String>> manufModelVersSerial = memoize(
-            WindowsBaseboardJNA::queryManufModelVersSerial);
+final class WindowsBaseboardJNA extends WindowsBaseboard {
 
     @Override
-    public String getManufacturer() {
-        return manufModelVersSerial.get().getA();
-    }
-
-    @Override
-    public String getModel() {
-        return manufModelVersSerial.get().getB();
-    }
-
-    @Override
-    public String getVersion() {
-        return manufModelVersSerial.get().getC();
-    }
-
-    @Override
-    public String getSerialNumber() {
-        return manufModelVersSerial.get().getD();
-    }
-
-    private static Quartet<String, String, String, String> queryManufModelVersSerial() {
-        String manufacturer = null;
-        String model = null;
-        String version = null;
-        String serialNumber = null;
-        WmiResult<BaseBoardProperty> win32BaseBoard = Win32BaseBoardJNA.queryBaseboardInfo();
-        if (win32BaseBoard.getResultCount() > 0) {
-            manufacturer = WmiUtil.getString(win32BaseBoard, BaseBoardProperty.MANUFACTURER, 0);
-            model = WmiUtil.getString(win32BaseBoard, BaseBoardProperty.MODEL, 0);
-            String product = WmiUtil.getString(win32BaseBoard, BaseBoardProperty.PRODUCT, 0);
-            if (!Util.isBlank(product)) {
-                model = Util.isBlank(model) ? product : (model + " (" + product + ")");
-            }
-            version = WmiUtil.getString(win32BaseBoard, BaseBoardProperty.VERSION, 0);
-            serialNumber = WmiUtil.getString(win32BaseBoard, BaseBoardProperty.SERIALNUMBER, 0);
-        }
-        return new Quartet<>(Util.isBlank(manufacturer) ? Constants.UNKNOWN : manufacturer,
-                Util.isBlank(model) ? Constants.UNKNOWN : model, Util.isBlank(version) ? Constants.UNKNOWN : version,
-                Util.isBlank(serialNumber) ? Constants.UNKNOWN : serialNumber);
+    protected WmiQueryExecutor getWmiQueryExecutor() {
+        return Objects.requireNonNull(WmiQueryExecutorJNA.createInstance());
     }
 }
